+++ address/Cargo.toml
@@ -9,6 +9,9 @@ license = "GPL-3.0"
 varint = { package = "unsigned-varint", version = "0.2.3" }
 blake2-rfc = { package = "blake2-rfc", version = "0.2.18" }
 data-encoding = { package = "data-encoding", version = "2.0.0" }
-serde = { version = "1.0", features = ["derive"] }
+bls = { package = "bls-signatures", version = "0.3" }
 libsecp256k1 = { version = "0.3.0" }
+serde = { version = "1.0", features = ["derive"] }
 derive_more = { version = "0.99.2"}
+rand = "0.7.2"
+hex = "0.4.0"
