+++ wallet/Cargo.toml
@@ -9,10 +9,8 @@ license = "GPL-3.0"
 bls = { package = "bls-signatures", version = "0.3" }
 libsecp256k1 = { version = "0.3.0" }
 derive_more = { version = "0.99.2"}
-hex = "0.4.0"
-rand = "0.7.2"
 serde_json = "1.0.41"
 parking_lot = "0.9.0"
-failure = "0.1.6"
+hex = "0.4.0"
 
 address = {path = "../address"}
